change id type to ulong

+++ Models/JsonIndexDocument.cs
@@ -7,7 +7,7 @@ namespace DoTuna
 #pragma warning disable IDE1006 // Naming Styles: Json has camelCase but c# is PascalCase
         public string version { get; set; } = "";
         public string boardId { get; set; } = "";
-        public int threadId { get; set; } = 0;
+        public ulong threadId { get; set; } = 0;
         public string title { get; set; } = "";
         public string username { get; set; } = "";
         public DateTime createdAt { get; set; } = new DateTime(0);
