+++ Services/ContentConverter.cs
@@ -19,7 +19,7 @@ namespace DoTuna
         }
 
         // >>threadId>>n 앵커 변환
-        protected string ConvertAnchors(string content, int threadId)
+        protected string ConvertAnchors(string content, ulong threadId)
         {
             return Regex.Replace(
                 content,
@@ -29,7 +29,7 @@ namespace DoTuna
                     var responseStart = m.Groups[3].Value;
                     if (string.IsNullOrEmpty(id) && string.IsNullOrEmpty(responseStart))
                         return m.Value;
-                    return MakeAnchorTag(int.Parse(id), responseStart, m.Value, false);
+                    return MakeAnchorTag(ulong.Parse(id), responseStart, m.Value, false);
                 },
                 RegexOptions.IgnoreCase
             );
@@ -44,7 +44,7 @@ namespace DoTuna
                 m => {
                     var threadId = m.Groups[2].Value;
                     var resNo = m.Groups[3].Value;
-                    return MakeAnchorTag(int.Parse(threadId), resNo, m.Value, true);
+                    return MakeAnchorTag(ulong.Parse(threadId), resNo, m.Value, true);
                 },
                 RegexOptions.IgnoreCase
             );
@@ -59,7 +59,7 @@ namespace DoTuna
                 m => {
                     var threadId = m.Groups[2].Value;
                     var resNo = m.Groups[3].Value;
-                    return MakeAnchorTag(int.Parse(threadId), resNo, m.Value, true);
+                    return MakeAnchorTag(ulong.Parse(threadId), resNo, m.Value, true);
                 },
                 RegexOptions.IgnoreCase
             );
@@ -74,7 +74,7 @@ namespace DoTuna
                 m => {
                     var threadId = m.Groups[2].Value;
                     // resNo가 없는 경우 빈 문자열 전달
-                    return MakeAnchorTag(int.Parse(threadId), "", m.Value, true);
+                    return MakeAnchorTag(ulong.Parse(threadId), "", m.Value, true);
                 },
                 RegexOptions.IgnoreCase
             );
@@ -91,10 +91,10 @@ namespace DoTuna
             );
         }
 
-        protected string GetFileName(int threadId) => Uri
+        protected string GetFileName(ulong threadId) => Uri
             .EscapeDataString(_fileNameMap.GetFileName(threadId));
 
-        public abstract string ConvertContent(string content, int threadId);
-        protected abstract string MakeAnchorTag(int threadId, string resNo, string text, bool isExternal);
+        public abstract string ConvertContent(string content, ulong threadId);
+        protected abstract string MakeAnchorTag(ulong threadId, string resNo, string text, bool isExternal);
     }
 }
